*** empty log message ***

+++ magick/blob.c
@@ -1313,7 +1313,7 @@ Export size_t ReadBlob(Image *image,const size_t length,void *data)
       */
       count=Min(length,image->blob.length-image->blob.offset);
       if (count > 0)
-        memcpy(data,image->blob.data+image->blob.offset,count);
+        memcpy(data,image->blob.data+image->blob.offset,(size_t) count);
       image->blob.offset+=count;
       return(count);
     }
